Create Collatz Conjecture in Python
What is Collatz conjecture: 
Take any natural number n. If n is even, divide it by 2 to get n / 2. If n is odd, multiply it by 3 and add 1 to obtain 3n + 1. Repeat the process (which has been called "Half Or Triple Plus One").Continue this process till you get 1. The conjecture is that no matter what number you start with, you will always eventually reach 1.
As the problem is good, here is a working solution in Python.